+++ src/main/java/open/commons/spring/web/autoconfigure/configuration/AuthorizedObjectMessageConfigureConfiguration.java
@@ -34,9 +34,11 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.context.annotation.Bean;
+import org.springframework.core.env.Environment;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import open.commons.spring.web.config.AuthorizedObjectMessageConfigure;
+import open.commons.spring.web.handler.AuthorizedModelAndViewHandlerInterceptor;
 import open.commons.spring.web.jackson.AuthorizedObjectJackson2HttpMessageConverter;
 
 /**
@@ -70,8 +72,9 @@ public class AuthorizedObjectMessageConfigureConfiguration {
 
     @Bean
     @ConditionalOnBean(name = { AuthorizedObjectJackson2HttpMessageConverter.BEAN_QUALIFIER })
-    WebMvcConfigurer authorizedObjectMessageConfigure(
-            @Qualifier(AuthorizedObjectJackson2HttpMessageConverter.BEAN_QUALIFIER) @NotNull AuthorizedObjectJackson2HttpMessageConverter messageConverter) {
+    WebMvcConfigurer authorizedObjectMessageConfigure(Environment environment //
+            , @Qualifier(AuthorizedObjectJackson2HttpMessageConverter.BEAN_QUALIFIER) @NotNull AuthorizedObjectJackson2HttpMessageConverter messageConverter //
+            , AuthorizedModelAndViewHandlerInterceptor authorizedModelAndViewHandlerInterceptor) {
         WebMvcConfigurer configure = new AuthorizedObjectMessageConfigure(messageConverter);
         logger.info("[authorized-resources] authorized-object-message-configure={}", configure);
         return configure;