+++ src/main/java/open/commons/spring/web/handler/PostProcessingHandlerInterceptor.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2025 Park Jun-Hong (parkjunhong77@gmail.com)
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ *
+ * This file is generated under this project, "open-commons-spring-web".
+ *
+ * Date  : 2025. 9. 25. 오후 3:53:20
+ *
+ * Author: parkjunhong77@gmail.com
+ * 
+ */
+
+package open.commons.spring.web.handler;
+
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistration;
+
+/**
+ * 
+ * @since 2025. 9. 25.
+ * @version 0.8.0
+ * @author parkjunhong77@gmail.com
+ */
+public interface PostProcessingHandlerInterceptor extends HandlerInterceptor {
+    /**
+     * {@link InterceptorRegistration}에 추가된 이후에 진행할 작업을 구현합니다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 9. 25.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param registry
+     *
+     * @since 2025. 9. 25.
+     * @version 0.8.0
+     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     */
+    default void afterRegistered(InterceptorRegistration registry) {
+    }
+}