+++ src/main/java/open/commons/spring/web/utils/PathUtils.java
@@ -27,6 +27,8 @@
 package open.commons.spring.web.utils;
 
 import java.util.Collection;
+import java.util.function.Consumer;
+import java.util.function.Function;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
 
@@ -34,6 +36,7 @@ import javax.annotation.Nonnull;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.core.env.Environment;
 
 /**
  * 
@@ -46,6 +49,69 @@ public class PathUtils {
     private static final Logger logger = LoggerFactory.getLogger(PathUtils.class);
 
     private PathUtils() {
+    }
+
+    /**
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 9. 25.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param env
+     *            환경설정 정보
+     * @param propName
+     *            설정이름
+     * @param actor
+     *            설정값을 사용할 대상.
+     *
+     * @since 2025. 9. 25.
+     * @version 0.8.0
+     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     */
+    public static void addEnvironmentProperty(@Nonnull Environment env, String propName, @Nonnull Consumer<String> actor) {
+        String value = env.getProperty(propName);
+        if (value == null) {
+            logger.debug("'{}' 에 해당하는 설정이 존재하지 않습니다.", propName);
+            return;
+        }
+        actor.accept(value);
+    }
+
+    /**
+     * 
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 9. 25.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param <V>
+     *            변환 데이터 유형
+     * @param env
+     *            환경설정 정보
+     * @param propName
+     *            설정이름
+     * @param converter
+     *            설정값 변환 함수
+     * @param actor
+     *            변환된 값을 사용할 대상.
+     *
+     * @since 2025. 9. 25.
+     * @version 0.8.0
+     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     */
+    public static <V> void addEnvironmentProperty(@Nonnull Environment env, String propName, Function<String, V> converter, @Nonnull Consumer<V> actor) {
+        String value = env.getProperty(propName);
+        if (value == null) {
+            logger.debug("'{}' 에 해당하는 설정이 존재하지 않습니다.", propName);
+            return;
+        }
+        actor.accept(converter.apply(value));
     }
 
     /**
@@ -123,5 +189,4 @@ public class PathUtils {
             return false;
         }
     }
-
 }